Move CompilerEnvState to package top level.
This splits the CompilerEnvState definition out of compiler_env.py and
into a top level package module.
