+++ compiler_gym/envs/__init__.py
@@ -2,19 +2,12 @@
 #
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
-from compiler_gym.envs.compiler_env import (
-    CompilerEnv,
-    CompilerEnvState,
-    info_t,
-    observation_t,
-    step_t,
-)
+from compiler_gym.envs.compiler_env import CompilerEnv, info_t, observation_t, step_t
 from compiler_gym.envs.llvm.llvm_env import LlvmEnv
 from compiler_gym.util.registration import COMPILER_GYM_ENVS
 
 __all__ = [
     "CompilerEnv",
-    "CompilerEnvState",
     "LlvmEnv",
     "observation_t",
     "info_t",
